// src/components/User/types.ts
export interface IFormProps {
    name: string
    email: string
    addressSuite: string
    addressStreet: string
    addressCity: string
    addressZipcode: string
}
export interface IFormErrorProps {
    name?: string
    email?: string
    addressSuite?: string
    addressStreet?: string
    addressCity?: string
    addressZipcode?: string
}